ci(e2e): recalibrate from 32-node runs, split merge-lanes-edge-cases, e2e_test parallelism 40
