+++ .circleci/config.yml
@@ -768,7 +768,7 @@ jobs:
     environment:
       BITSRC_ENV: stg
       BIT_FEATURES: cloud-importer-v2
-    parallelism: 32
+    parallelism: 40
     steps:
       - e2e_test_cmd
